Generating path
Now able to fly above, generate path, then follow it. Commands are all done manually so probably need to make a more streamlined way to do it. Also, path has an arbitrary offset so that drone doesn't collide with the object, but will need to be adjusted for real applications.