added spiral path
Able to generate a spiral path now, just need to change the function used in the read_stereo cpp file depending on which one you want. Path following almost seems to get worse with slower speed so im not sure what is going on thre but it is currently set very slow.